refactor(settings): move ChatGPT subscription to the Account tab
The subscription is an account connection, not a model setting: it now
sits with the TeXRA sign-in and above the provider credentials it stands
in for. Telemetry moves below credentials on the same page.

The section moves out of ModelsTab into a CodexSubscriptionSection
component, named after the existing codexSubscriptionSignIn convention.
