+++ packages/extension/src/settingsView/frontend/tabs/ModelsTab.ts
@@ -3,9 +3,6 @@
 import { LitElement, html, nothing, css, type TemplateResult } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-// Local imports - shared auth
-import { codexAccountLabel } from '@auth/codex/codexSessionTypes';
-
 // Local imports - shared styles
 import { commonViewStyles, designTokens } from '@shared/styles';
 import { SETTINGS_VIEW_COMMANDS } from '@shared/ipc';
@@ -17,12 +14,10 @@ import { waIcon } from '@shared/wa/webAwesomeIcons';
 // Web Awesome icon bundle (side-effect import)
 import '@awesome.me/webawesome/dist/components/button/button.js';
 import '@awesome.me/webawesome/dist/components/icon/icon.js';
-import '@awesome.me/webawesome/dist/components/switch/switch.js';
 import '@awesome.me/webawesome/dist/components/tag/tag.js';
 
 // Local imports - shared schemas
 import type {
-  ChatGptAuthStatus,
   CopilotRouteInfo,
   ModelSelectionItem,
   NumberSetting,
@@ -37,7 +32,6 @@ import '../components/profile/ApiAccessSection';
 import '../components/profile/ProviderKeyList';
 import '../components/profile/ModelSelectionList';
 import '../components/profile/ReliabilitySettingsSection';
-import type WaSwitch from '@awesome.me/webawesome/dist/components/switch/switch.js';
 
 @customElement('models-tab')
 export class ModelsTab extends LitElement {
@@ -82,7 +76,6 @@ export class ModelsTab extends LitElement {
   @property({ attribute: false }) apiAccessMode: 'included' | 'personal' =
     'personal';
   @property({ attribute: false }) providerKeyStatuses: ProviderKeyStatus[] = [];
-  @property({ attribute: false }) chatgptAuth: ChatGptAuthStatus | null = null;
   @property({ attribute: false }) globalStreamingDefault = true;
   @property({ attribute: false }) modelSelectionItems: ModelSelectionItem[] =
     [];
@@ -90,13 +83,6 @@ export class ModelsTab extends LitElement {
   @property({ attribute: false }) reliabilitySettings: NumberSetting[] = [];
   @property({ attribute: false }) helperModel = '';
   @property({ type: Boolean }) preferShortModelNames = false;
-
-  private readonly handlePreferSubscriptionChange = (event: Event): void => {
-    const enabled = (event.target as WaSwitch).checked;
-    postMessage(SETTINGS_VIEW_COMMANDS.SET_CHATGPT_PREFER_SUBSCRIPTION, {
-      enabled,
-    });
-  };
 
   override render(): TemplateResult {
     const apiAccessSection = this.authenticated
@@ -107,8 +93,7 @@ export class ModelsTab extends LitElement {
 
     return html`
       <div class="models-container tab-content-container">
-        ${apiAccessSection} ${this.renderChatGptSection()}
-        ${this.renderCopilotSection()}
+        ${apiAccessSection} ${this.renderCopilotSection()}
         <provider-key-list
           .providerKeyStatuses=${this.providerKeyStatuses}
           .apiAccessMode=${this.apiAccessMode}
@@ -124,92 +109,6 @@ export class ModelsTab extends LitElement {
           .settings=${this.reliabilitySettings}
         ></reliability-settings-section>
       </div>
-    `;
-  }
-
-  /**
-   * Experimental "Sign in with ChatGPT" control. After sign-in, the configured
-   * Codex-eligible models run on the user's own ChatGPT subscription (when
-   * `chatgptCodex.preferSubscription` is enabled) instead of an API key.
-   */
-  private renderChatGptSection(): TemplateResult {
-    const signedIn = this.chatgptAuth?.signedIn ?? false;
-    const preferSubscription = this.chatgptAuth?.preferSubscription ?? false;
-    const account = codexAccountLabel(this.chatgptAuth);
-    return html`
-      <section id="chatgpt-subscription">
-        ${renderSettingsSectionHeading({
-          title: 'ChatGPT subscription',
-          description:
-            'Use OpenAI models through your ChatGPT Plus, Pro, or Team subscription. No OpenAI API key is needed.',
-          actions: html`<wa-tag variant="neutral" size="s"
-            >Experimental</wa-tag
-          >`,
-        })}
-        <p class="keyless-source__limit">
-          ${waIcon('circle-info')}
-          <span>
-            Subscription routing currently uses a 272,000-token Codex context
-            cap, not the full 1,000,000-token API context.
-          </span>
-        </p>
-        <div class="settings-section">
-          <div class="settings-row">
-            <div class="settings-row-text">
-              <span class="settings-row-label">
-                Prefer ChatGPT subscription
-              </span>
-              <span class="settings-row-help">
-                Route eligible Codex models through your subscription.
-              </span>
-            </div>
-            <wa-switch
-              class="settings-row-control"
-              aria-label="Prefer ChatGPT subscription"
-              ?checked=${preferSubscription}
-              @change=${this.handlePreferSubscriptionChange}
-            ></wa-switch>
-          </div>
-          <div class="settings-row">
-            <div class="settings-row-text">
-              <span class="settings-row-label">
-                ${
-                  signedIn
-                    ? html`${waIcon('circle-check')} Signed in as ${account}`
-                    : 'ChatGPT account'
-                }
-              </span>
-              <span class="settings-row-help">
-                ${
-                  signedIn
-                    ? 'Your subscription is ready to use.'
-                    : 'Connect the ChatGPT account that owns your subscription.'
-                }
-              </span>
-            </div>
-            <div class="settings-row-control">
-              ${
-                signedIn
-                  ? renderLabeledActionButton({
-                      text: 'Sign out',
-                      kind: 'secondary',
-                      appearance: 'outlined',
-                      onClick: () =>
-                        postMessage(SETTINGS_VIEW_COMMANDS.SIGN_OUT_CHATGPT),
-                    })
-                  : renderLabeledActionButton({
-                      icon: 'comments',
-                      text: 'Sign in with ChatGPT',
-                      kind: 'primary',
-                      appearance: 'filled',
-                      onClick: () =>
-                        postMessage(SETTINGS_VIEW_COMMANDS.SIGN_IN_CHATGPT),
-                    })
-              }
-            </div>
-          </div>
-        </div>
-      </section>
     `;
   }
